chore(deps): upgrade 14 workspace dependencies to latest
Batch 1 — safe updates:
- rusqlite 0.33 → 0.38 (u64/usize → i64 casts for FromSql/ToSql)
- git2 0.18 → 0.20 (+ vendored sem-core)
- toml 0.9.5 → 1.0
- agent-client-protocol 0.9.4 → 0.10 (FileSystemCapability → FileSystemCapabilities)
- scraper 0.20 → 0.25
- sysinfo 0.30 → 0.38
- wiremock 0.5 → 0.6
- chromiumoxide 0.7 → 0.9

Batch 2 — medium breaking:
- thiserror 1 → 2
- rand 0.8 → 0.9 (thread_rng → rng, gen_range → random_range)
- jsonschema 0.18 → 0.28 (JSONSchema::compile → validator_for)

Batch 3 — major rewrites:
- rmcp 0.15 → 1.2 (non-exhaustive structs → builder pattern)
- reqwest 0.12 → 0.13 (rustls-tls → rustls feature)
- serde_yaml 0.9 → serde_yaml_ng 0.10 (deprecated crate replaced)

Also converted 8 inline version pins to workspace = true in:
swissarmyhammer-lsp, model-embedding, swissarmyhammer-embedding, ane-embedding

+++ swissarmyhammer-code-context/src/ops/find_duplicates.rs
@@ -255,7 +255,7 @@ mod tests {
         conn.execute(
             "INSERT INTO ts_chunks (file_path, start_byte, end_byte, start_line, end_line, symbol_path, text, embedding)
              VALUES (?1, 0, ?2, ?3, ?4, ?5, ?6, ?7)",
-            rusqlite::params![file_path, text.len(), start_line, end_line, symbol_path, text, blob],
+            rusqlite::params![file_path, text.len() as i64, start_line, end_line, symbol_path, text, blob],
         )
         .unwrap();
     }
